+++ lib/cocoapods/config.rb
@@ -314,6 +314,18 @@ module Pod
       nil
     end
 
+    # Excludes the given dir from Time Machine backups.
+    #
+    # @param  [Pathname] dir
+    #         The directory to exclude from Time Machine backups.
+    #
+    # @return [void]
+    #
+    def exclude_from_backup(dir)
+      return if Gem.win_platform?
+      system('tmutil', 'addexclusion', dir.to_s, %i(out err) => File::NULL)
+    end
+
     public
 
     #-------------------------------------------------------------------------#